meg28_birthday_bot/requirements.txt
pandas
Pillow
selenium
webdriver-manager
requests